Updated facemesh demo to SDK, moved api implementation to new folder
