+++ gen2-facemesh/requirements.txt
@@ -1,3 +1,3 @@
 opencv-python
 numpy
-depthai==2.16.0.0
+depthai-sdk==1.9.6
